+++ ragframework/document/__init__.py
@@ -2,12 +2,13 @@
 
 from ragframework.document.chunkers import FixedSizeChunker, RecursiveChunker, SentenceChunker
 
-from .loaders import MarkdownLoader, PDFLoader, TextFileLoader
+from .loaders import DocxLoader, MarkdownLoader, PDFLoader, TextFileLoader
 
 __all__ = [
     "TextFileLoader",
     "MarkdownLoader",
     "PDFLoader",
+    "DocxLoader",
     "FixedSizeChunker",
     "RecursiveChunker",
     "SentenceChunker",