+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `DocxLoader` with support for per-paragraph and whole-file modes (closes #2)
+
 ## [0.2.0] - 2026-09-19
 
 ### Added